Introduce Data.Aeson.Text and tidy up modules a bit
