--- Data/Aeson/Encode/Functions.hs
@@ -1,14 +1,0 @@
-{-# LANGUAGE CPP #-}
-
-module Data.Aeson.Encode.Functions
-    (
-      builder
-    ) where
-
-import Data.Aeson.Encoding
-import Data.Aeson.Types.Class
-import Data.ByteString.Builder (Builder)
-import Data.Monoid ((<>))
-
-
-{-# INLINE builder #-}
