+++ benchmarks/Micro.hs
@@ -1,7 +1,7 @@
 module Main (main) where
 
 import Criterion.Main
-import qualified Data.Aeson.Encode.Builder as AB
+import qualified Data.Aeson.Encoding.Builder as AB
 import qualified Data.ByteString.Builder as B
 import qualified Data.Text as T
