+++ benchmarks/aeson-benchmarks.cabal
@@ -9,17 +9,18 @@ library
 
   exposed-modules:
     Data.Aeson
-    Data.Aeson.Encode
     Data.Aeson.Encoding
-    Data.Aeson.Encoding.Internal
-    Data.Aeson.Internal
-    Data.Aeson.Internal.Time
     Data.Aeson.Parser
+    Data.Aeson.Text
     Data.Aeson.Types
     Data.Aeson.TH
 
-    Data.Aeson.Encode.Builder
-    Data.Aeson.Functions
+    Data.Aeson.Encoding.Internal
+    Data.Aeson.Internal
+    Data.Aeson.Internal.Time
+
+    Data.Aeson.Encoding.Builder
+    Data.Aeson.Internal.Functions
     Data.Aeson.Parser.Internal
     Data.Aeson.Parser.Time
     Data.Aeson.Types.FromJSON
